Move function to gatsby-browser
